Move auth modules into src/auth and merge oauthGrant into useOAuthFlow
The auth subsystem (authStore, requestAccessToken, useOAuthFlow,
useChromaticDialog, subdomain) had been spread across src/utils with no
visible feature boundary, and oauthGrant.ts was a thin shim that only
useOAuthFlow consumed. Newcomers had to rebuild the OAuth picture from
unrelated utility files.

Relocate the auth files into src/auth so the feature reads as a unit,
inline parseGrantPayload and exchangeOAuthCode into useOAuthFlow (its
sole consumer), and update every import + test mock to the new paths.
graphQLClient stays in utils but now imports authStore / AuthSession
from src/auth.
